farmerapp/project/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
flask-cors
Flask-JWT-Extended
bcrypt